Add deepcopy generation to manifests generation task
This commit adds a generate_deepcopy function to the tasks.py file,
simplifying the process of generating deepcopy functions when updating
CRDs. The function is automatically invoked by the generate-manifests
task, which is already triggered during CRD updates. The manifests are
generated using controller-gen, and the Apache License is added at the
beginning of the generated code.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>